--- nuclio/benchmark/tsdb-bench-test-config.yaml
@@ -1,28 +1,0 @@
-# File: tsdb-bench-test-config.yaml
-
-# Control test verbosity (in Nuclio only)
-Verbose: false
-
-# Relative time, i.e. now() - startTimeOffset
-StartTimeOffset: "72h"
-
-# Interval in milliseconds between samples
-SampleStepSize: 1000
-
-# should be in range [1..26], i.e [A..Z]
-NamesCount: 26
-
-# Pattern: Name_[A..Z][_[1..200]]
-NamesDiversity: 10
-
-# should be in range [1..26]
-LabelsCount: 3
-
-# Pattern: Label_[A..Z][_[1..10]]
-LabelsDiversity: 1
-
-# should be in range [1..26], i.e [A..Z]
-LabelValuesCount: 26
-
-# Pattern: [A..Z][_[1..10]]
-LabelsValueDiversity: 10
